Require 200 Energy To Spawn A Worker
When a room collapses you can only safely assume that there is 0 energy in the extension and 0 energy in the spawn. The game will give you a full spawn of energy after a while and from this life can resume.  We should always be able to spawn a worker to drive room recovery with that 200 energy.  

This is a simple fix for issue #614 Must Always Be Able To Build A Worker With 200 Energy.  

It could be improved by making the minimum energy level required to spawn a worker dependent on the population of the room.  If there are other workers or haulers in the room, we could use the higher level.  This would ensure larger workers are always made at higher levels unless a population collapse has happened.  The advantage of this change is that it is simple, it works and it solves the issue this second.